Feat: shopId, noticeId 쿼리로 받아오기
